+++ internal/resources/workflow_replication/resource.go
@@ -266,7 +266,7 @@ func (r *WorkflowReplicationResource) resolveInstanceID(model *WorkflowReplicati
 func modelToAPIReplication(model *WorkflowReplicationResourceModel) *client.ReplicationStream {
 	stream := &client.ReplicationStream{
 		Name:    model.Name.ValueString(),
-		Version: int(model.Version.ValueInt64()),
+		Version: model.Version.ValueInt64(),
 		Enabled: model.Enabled.ValueBool(),
 	}
 
@@ -313,7 +313,7 @@ func apiReplicationToModel(stream *client.ReplicationStream, model *WorkflowRepl
 		model.WorkflowID = types.StringValue(stream.StreamID)
 	}
 	model.Name = types.StringValue(stream.Name)
-	model.Version = types.Int64Value(int64(stream.Version))
+	model.Version = types.Int64Value(stream.Version)
 	model.Enabled = types.BoolValue(stream.Enabled)
 
 	if stream.Source != nil {
